visualizer/pyproject.toml
[project]
name = "rag-visualizer"
version = "0.1.0"
description = "Export ChromaDB embeddings to TensorBoard Projector"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.6.0",
    "tensorboard>=2.15.0",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.1.0",
    # tensorboard imports `pkg_resources` from setuptools at runtime.
    # setuptools 81+ removed pkg_resources, so pin below it.
    "setuptools>=70.0.0,<81",
]
